Functional.
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/internal/pkgtools@15069 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ release-binary-packages.sh
@@ -6,22 +6,27 @@ usage() {
   exit 1
 }
 
+MAX_DEPTH="-maxdepth 1"
+
 while getopts "r:d:ah?" opt ; do
   case "$opt" in
     r) REPOSITORY=$OPTARG ;;
     d) DISTRIBUTION=$OPTARG ;;
-    a) MAX_DEPTH="-maxdepth 1" ;;
+    a) MAX_DEPTH="" ;;
     h|\?) usage ;;
   esac
 done
 
 [ -z "$REPOSITORY" ] || [ -z "$DISTRIBUTION" ] && usage
 
-DEBS=$(find . $MAXDEPTH -name "*.deb")
+DEBS=$(find . $MAX_DEPTH -name "*.deb" | xargs)
 
-for p in $DEBS ; do
-  touch ${p/.deb/.$REPOSITORY_$DISTRIBUTION.manifest}
-  MANIFESTS=$(find . $MAXDEPTH -name "*.manifest")
+if [ -n "$DEBS" ] ; then
+  for p in $DEBS ; do
+    touch ${p/.deb/.$REPOSITORY_$DISTRIBUTION.manifest}
+  done
+
+  MANIFESTS=$(find . $MAXDEPTH -name "*.manifest" | xargs)
   lftp -e "set net:max-retries 1 ; cd incoming ; put $DEBS $MANIFESTS ; exit" mephisto
   rm -f $MANIFESTS
-done
+fi
